# src/backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "家庭饮食管理系统的后端API接口文档",
        "title": "OneTaste Family API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.onetastefamily.com/support",
            "email": "support@onetastefamily.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/captcha": {
            "get": {
                "description": "获取一个带干扰的图形验证码图片及对应的编码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "获取图形验证码",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "图片宽度（120-360），默认220",
                        "name": "width",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "图片高度（40-160），默认70",
                        "name": "height",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证码获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.CaptchaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "用户登录接口，支持手机号和密码登录。登录成功后返回用户ID、JWT Token和过期时间。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "手机号或密码错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "用户注册接口，支持手机号、密码、验证码、昵称注册。注册成功后返回用户ID和JWT Token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.RegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误、验证码错误或手机号已注册",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/media/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "上传图片等媒体文件，并返回可访问地址。需要 Bearer Token。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "媒体"
                ],
                "summary": "上传媒体文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "待上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件目录前缀，例如 /user/head/{userId}、/family/{familyId}/dishes/{dishId}",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.MediaUploadResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CaptchaResponse": {
            "description": "图形验证码返回数据",
            "type": "object",
            "properties": {
                "captcha_key": {
                    "description": "验证码编码",
                    "type": "string",
                    "example": "01J0XYZABCD1234EFG567HIJK"
                },
                "expire_in": {
                    "description": "过期时间（秒）",
                    "type": "integer",
                    "example": 30
                },
                "image_base64": {
                    "description": "图形验证码Base64",
                    "type": "string",
                    "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUg..."
                }
            }
        },
...
        "models.LoginRequest": {
            "description": "用户登录请求参数",
            "type": "object",
            "required": [
                "password",
                "phone"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "password123"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string",
                    "example": "13800138000"
                }
            }
        },
        "models.LoginResponse": {
            "description": "用户登录成功返回的数据",
            "type": "object",
            "properties": {
                "expires_in": {
                    "description": "Token过期时间（秒）",
                    "type": "integer",
                    "example": 86400
                },
                "token": {
                    "description": "JWT Token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string",
                    "example": "01HZX1YF8Y6S7K4V9Q2J3M5N6P"
                }
            }
        },
        "models.MediaUploadResponse": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "directory": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "object_key": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.RegisterRequest": {
            "description": "用户注册请求参数",
            "type": "object",
            "required": [
                "captcha_key",
                "nickname",
                "password",
                "phone",
                "verify_code"
            ],
            "properties": {
                "captcha_key": {
                    "description": "验证码编码",
                    "type": "string",
                    "example": "01J0XYZABCD1234EFG567HIJK"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string",
                    "example": "张三"
                },
                "password": {
                    "description": "密码，至少6位",
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "phone": {
                    "description": "手机号，11位数字",
                    "type": "string",
                    "example": "13800138000"
                },
                "verify_code": {
                    "description": "图形验证码内容，不区分大小写",
                    "type": "string",
                    "example": "A9d3"
                }
            }
        },
        "models.RegisterResponse": {
            "description": "用户注册成功返回的数据",
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT Token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string",
                    "example": "01HZX1YF8Y6S7K4V9Q2J3M5N6P"
                }
            }
        },
...
        "utils.FieldError": {
            "description": "字段验证错误信息",
            "type": "object",
            "properties": {
                "field": {
                    "description": "错误字段名",
                    "type": "string",
                    "example": "phone"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "手机号格式不正确"
                }
            }
        },
        "utils.Response": {
            "description": "统一的API响应格式",
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据"
                },
                "errors": {
                    "description": "字段错误列表（仅在错误时返回）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.FieldError"
                    }
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 \"Bearer {token}\" 格式，token 通过登录接口获取",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}